requirements.txt
Django==1.11.2
django-tagging